MultiAgent_DocStudio/requirements.txt
streamlit==1.28.1
openai==1.3.7
pymupdf==1.23.8
pandas==2.1.3
python-docx==1.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
tiktoken==0.5.1 